+++ docs/CHANGELOG.rst
@@ -16,6 +16,7 @@ Added
 - Add set_permission method on PermissionQuerySet
 - Add notify_create method on Subscription model in observers to enable
   sending notification when object is created
+- Allow ordering data endpoint by entity name
 
 Changed
 -------